feat: Get Balance
In this commit, i have implemented "Get Balance" feature in the Balance
group according to official documentation.

To accomplish it, i have added a method called get_with_params where the
method is used if there are parameters in the GET operation.

After that, i have installed 'serde_urlencoded', 'data-encoding', and
'dotenv' where 'serde_urlencoded' is used to perform encoding of
parameters. Then 'data-encoding' is used to encode Xendit API Key using
base64 string following Basic Authentication Standart in RFC 7617.  And
'dotenv' is used in dev-dependencies and used for testing.

Signed-off-by: Taufiq Pohan <m.taufiq30s@hotmail.com>