fix: Implement Box<dyn Error> to prevent panic
In this commit, i have implemented Box<dyn Error> to prevent
applications that use this library from force closing when an error
occurs from the library.

The reason i use Box<dyn Error> is because there are several error
types, such as serde::qs_Error and reqwest::Error and also currently i
am focusing more on the error message than the error type. There may be
improvements in the future to make error handling in this library more
better.

Signed-off-by: Taufiq Pohan <m.taufiq30s@hotmail.com>