fix: Missing properties in InvoiceResponse
The field 'identity_amount' was missing from 'Bank' struct. This causes
the deserialize process to fail. To solve this problem, add field
'identity_amount' to 'Bank' struct.

In addiction, change the data type of 'transfer_amount' to 'Option<f64>'
because this field will be depracted soon.

And update the unit test for invoice by clariflying the unit test
function name and modifying the 'test_get_invoice' unit test by add code
to create new invoice and retrieving it.

Signed-off-by: Taufiq Pohan <m.taufiq30s@hotmail.com>